+++ src/RcsbGraphQL/RcsbClient.ts
@@ -7,13 +7,14 @@ import {
     QueryAnnotationsArgs
 } from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
 import {
-    CoreEntry, CorePolymerEntity, QueryEntriesArgs,
-    QueryEntryArgs, QueryPolymer_EntitiesArgs
+    CoreEntry, CoreInterface, CorePolymerEntity, QueryEntriesArgs,
+    QueryEntryArgs, QueryInterfacesArgs, QueryPolymer_EntitiesArgs
 } from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
 import {RcsbQueryEntryInstances} from "./RcsbQueryEntryInstances";
 import {RcsbQueryMultipleEntityInstances} from "./RcsbQueryMultipleEntityInstances";
 import {RcsbCoreQueryInterface} from "./RcsbCoreQueryInterface";
 import {RcsbQueryMultipleEntriesProperties} from "./RcsbQueryMultipleEntriesProperties";
+import {RcsbQueryInterfaceInstances} from "./RcsbQueryInterfaceInstances";
 
 export class RcsbClient {
     private rcsbQueryAnnotations: RcsbCoreQueryInterface<QueryAnnotationsArgs,Array<AnnotationFeatures>> = new RcsbQueryAnnotations();
@@ -21,6 +22,7 @@ export class RcsbClient {
     private rcsbQueryEntityInstances: RcsbCoreQueryInterface<QueryEntryArgs,CoreEntry> = new RcsbQueryEntryInstances();
     private rcsbQueryMutipleEntityInstances: RcsbCoreQueryInterface<QueryPolymer_EntitiesArgs,Array<CorePolymerEntity>> = new RcsbQueryMultipleEntityInstances();
     private rcsbQueryEntryProperties: RcsbCoreQueryInterface<QueryEntriesArgs,Array<CoreEntry>> = new RcsbQueryMultipleEntriesProperties();
+    private rcsbQueryInterfaceInstances: RcsbCoreQueryInterface<QueryInterfacesArgs,Array<CoreInterface>> = new RcsbQueryInterfaceInstances();
 
     public async requestRcsbPdbAnnotations(requestConfig: QueryAnnotationsArgs): Promise<Array<AnnotationFeatures>>{
         return await this.rcsbQueryAnnotations.request(requestConfig);
@@ -40,5 +42,9 @@ export class RcsbClient {
 
     public async requestMultipleEntriesProperties(requestConfig: QueryEntriesArgs): Promise<Array<CoreEntry>>{
         return await this.rcsbQueryEntryProperties.request(requestConfig);
+    }
+
+    public async requestInterfaceInstances(requestConfig:QueryInterfacesArgs): Promise<Array<CoreInterface>> {
+        return await this.rcsbQueryInterfaceInstances.request(requestConfig);
     }
 }